put crazy style back
