+++ example/lib/main.dart
@@ -3,149 +3,11 @@
 import 'dart:ui';
 
 import 'package:arborio/tree_view.dart';
+import 'package:arborio_sample/data.dart';
+import 'package:arborio_sample/functions.dart';
 import 'package:arborio_sample/image_builder.dart';
 import 'package:flutter/material.dart';
 import 'package:path/path.dart' as path;
-
-enum ElementType { file, folder }
-
-class FileSystemElement {
-  FileSystemElement(this.name, this.type);
-
-  final String name;
-  final ElementType type;
-}
-
-// Initialize your tree nodes with FileSystemElement type
-List<TreeNode<FileSystemElement>> fileTree() => [
-      TreeNode<FileSystemElement>(
-        const Key('Projects'),
-        FileSystemElement('Projects', ElementType.folder),
-        [
-          TreeNode<FileSystemElement>(
-            const Key('FlutterApp'),
-            FileSystemElement('FlutterApp', ElementType.folder),
-            [
-              TreeNode<FileSystemElement>(
-                const Key('lib'),
-                FileSystemElement('lib', ElementType.folder),
-                [
-                  TreeNode<FileSystemElement>(
-                    const Key('main.dart'),
-                    FileSystemElement('main.dart', ElementType.file),
-                  ),
-                  TreeNode<FileSystemElement>(
-                    const Key('app.dart'),
-                    FileSystemElement('app.dart', ElementType.file),
-                  ),
-                ],
-              ),
-              TreeNode<FileSystemElement>(
-                const Key('assets'),
-                FileSystemElement('assets', ElementType.folder),
-                [
-                  TreeNode<FileSystemElement>(
-                    const Key('logo.png'),
-                    FileSystemElement('logo.png', ElementType.file),
-                  ),
-                  TreeNode<FileSystemElement>(
-                    const Key('data.json'),
-                    FileSystemElement('data.json', ElementType.file),
-                  ),
-                ],
-              ),
-            ],
-          ),
-          TreeNode<FileSystemElement>(
-            const Key('PythonScripts'),
-            FileSystemElement('PythonScripts', ElementType.folder),
-            [
-              TreeNode<FileSystemElement>(
-                const Key('script.py'),
-                FileSystemElement('script.py', ElementType.file),
-              ),
-            ],
-          ),
-        ],
-      ),
-      TreeNode<FileSystemElement>(
-        const Key('Documents'),
-        FileSystemElement('Documents', ElementType.folder),
-        [
-          TreeNode<FileSystemElement>(
-            const Key('Resume.docx'),
-            FileSystemElement('Resume.docx', ElementType.file),
-          ),
-          TreeNode<FileSystemElement>(
-            const Key('Budget.xlsx'),
-            FileSystemElement('Budget.xlsx', ElementType.file),
-          ),
-        ],
-      ),
-      TreeNode<FileSystemElement>(
-        const Key('Music'),
-        FileSystemElement('Music', ElementType.folder),
-        [
-          TreeNode<FileSystemElement>(
-            const Key('Favorites'),
-            FileSystemElement('Favorites', ElementType.folder),
-            [
-              TreeNode<FileSystemElement>(
-                const Key('song1.mp3'),
-                FileSystemElement('song1.mp3', ElementType.file),
-              ),
-              TreeNode<FileSystemElement>(
-                const Key('song2.mp3'),
-                FileSystemElement('song2.mp3', ElementType.file),
-              ),
-            ],
-          ),
-        ],
-      ),
-      TreeNode<FileSystemElement>(
-        const Key('Photos'),
-        FileSystemElement('Photos', ElementType.folder),
-        [
-          TreeNode<FileSystemElement>(
-            const Key('Vacation'),
-            FileSystemElement('Vacation', ElementType.folder),
-            [
-              TreeNode<FileSystemElement>(
-                const Key('image1.jpg'),
-                FileSystemElement('image1.jpg', ElementType.file),
-              ),
-              TreeNode<FileSystemElement>(
-                const Key('image2.jpg'),
-                FileSystemElement('image2.jpg', ElementType.file),
-              ),
-            ],
-          ),
-          TreeNode<FileSystemElement>(
-            const Key('Family'),
-            FileSystemElement('Family', ElementType.folder),
-            [
-              TreeNode<FileSystemElement>(
-                const Key('photo1.jpg'),
-                FileSystemElement('photo1.jpg', ElementType.file),
-              ),
-            ],
-          ),
-        ],
-      ),
-    ];
-
-const defaultExpander = Icon(Icons.chevron_right);
-
-const arrowRight = Icon(Icons.arrow_right);
-
-const doubleArrow = Icon(Icons.double_arrow);
-
-const fingerPointer = Text(
-  '👉',
-  style: TextStyle(fontSize: 16),
-);
-
-enum DisplayStyle { modern, retro }
 
 // The main app widget
 void main() {
@@ -167,7 +29,7 @@ class _MyAppState extends State<MyApp> {
   final textEditingController = TextEditingController(text: '500');
   TreeNode<FileSystemElement>? _selectedNode;
   final List<TreeNode<FileSystemElement>> _fileTree = fileTree();
-  DisplayStyle _displayStyle = DisplayStyle.modern;
+  DisplayStyle _displayStyle = DisplayStyle.crazy;
 
   @override
   Widget build(BuildContext context) => MaterialApp(
@@ -196,6 +58,7 @@ class _MyAppState extends State<MyApp> {
           ),
           body: Stack(
             children: [
+              // Background image
               Opacity(
                 opacity: .025,
                 child: imageAsset(
@@ -205,7 +68,9 @@ class _MyAppState extends State<MyApp> {
                   height: double.infinity,
                 ),
               ),
+              // Tree view
               _treeView(),
+              // Bottom pane
               Positioned(
                 bottom: 0,
                 left: 0,
@@ -217,6 +82,7 @@ class _MyAppState extends State<MyApp> {
         ),
       );
 
+  /// The bottom pane with the duration field, dropdowns, buttons, and style
   ClipRRect _bottomPane() => ClipRRect(
         child: BackdropFilter(
           filter: ImageFilter.blur(sigmaX: 5, sigmaY: 5),
@@ -243,6 +109,7 @@ class _MyAppState extends State<MyApp> {
         ),
       );
 
+  /// The duration field for setting the animation duration
   Widget _durationField() => Center(
         child: SizedBox(
           width: 200,
@@ -264,9 +131,11 @@ class _MyAppState extends State<MyApp> {
         ),
       );
 
+  /// The title of the app
   String _title() => 'Arborio Sample${_selectedNode != null ? ' - '
       '${_selectedNode!.data.name}' : ''}';
 
+  /// The row with the dropdowns for the expander and animation curve
   Widget _dropDownsRow() => Row(
         children: [
           DropdownMenu<Widget>(
@@ -290,69 +159,73 @@ class _MyAppState extends State<MyApp> {
             ],
           ),
           const SizedBox(width: 16),
-          DropdownMenu<String>(
-            label: const Text('Animation Curve'),
-            onSelected: (v) => _selectedCurve = v ?? _selectedCurve,
-            initialSelection: _selectedCurve,
-            dropdownMenuEntries: const [
-              DropdownMenuEntry(value: 'bounceIn', label: 'bounceIn'),
-              DropdownMenuEntry(value: 'bounceInOut', label: 'bounceInOut'),
-              DropdownMenuEntry(value: 'bounceOut', label: 'bounceOut'),
-              DropdownMenuEntry(value: 'ease', label: 'ease'),
-              DropdownMenuEntry(value: 'easeIn', label: 'easeIn'),
-              DropdownMenuEntry(value: 'easeInBack', label: 'easeInBack'),
-              DropdownMenuEntry(value: 'easeInCirc', label: 'easeInCirc'),
-              DropdownMenuEntry(value: 'easeInExpo', label: 'easeInExpo'),
-              DropdownMenuEntry(value: 'easeInOut', label: 'easeInOut'),
-              DropdownMenuEntry(
-                value: 'easeInOutBack',
-                label: 'easeInOutBack',
-              ),
-              DropdownMenuEntry(
-                value: 'easeInOutCirc',
-                label: 'easeInOutCirc',
-              ),
-              DropdownMenuEntry(
-                value: 'easeInOutExpo',
-                label: 'easeInOutExpo',
-              ),
-              DropdownMenuEntry(
-                value: 'easeInOutQuad',
-                label: 'easeInOutQuad',
-              ),
-              DropdownMenuEntry(
-                value: 'easeInOutQuart',
-                label: 'easeInOutQuart',
-              ),
-              DropdownMenuEntry(
-                value: 'easeInOutQuint',
-                label: 'easeInOutQuint',
-              ),
-              DropdownMenuEntry(
-                value: 'easeInOutSine',
-                label: 'easeInOutSine',
-              ),
-              DropdownMenuEntry(value: 'easeInQuad', label: 'easeInQuad'),
-              DropdownMenuEntry(value: 'easeInQuart', label: 'easeInQuart'),
-              DropdownMenuEntry(value: 'easeInQuint', label: 'easeInQuint'),
-              DropdownMenuEntry(value: 'easeInSine', label: 'easeInSine'),
-              DropdownMenuEntry(value: 'easeOut', label: 'easeOut'),
-              DropdownMenuEntry(value: 'easeOutBack', label: 'easeOutBack'),
-              DropdownMenuEntry(value: 'easeOutCirc', label: 'easeOutCirc'),
-              DropdownMenuEntry(value: 'easeOutExpo', label: 'easeOutExpo'),
-              DropdownMenuEntry(value: 'easeOutQuad', label: 'easeOutQuad'),
-              DropdownMenuEntry(value: 'easeOutQuart', label: 'easeOutQuart'),
-              DropdownMenuEntry(value: 'easeOutQuint', label: 'easeOutQuint'),
-              DropdownMenuEntry(value: 'easeOutSine', label: 'easeOutSine'),
-              DropdownMenuEntry(value: 'elasticIn', label: 'elasticIn'),
-              DropdownMenuEntry(value: 'elasticInOut', label: 'elasticInOut'),
-              DropdownMenuEntry(value: 'elasticOut', label: 'elasticOut'),
-              DropdownMenuEntry(value: 'linear', label: 'linear'),
-            ],
-          ),
+          _curveDropDown(),
         ],
       );
 
+  /// The dropdown for selecting the animation curve
+  DropdownMenu<String> _curveDropDown() => DropdownMenu<String>(
+        label: const Text('Animation Curve'),
+        onSelected: (v) => _selectedCurve = v ?? _selectedCurve,
+        initialSelection: _selectedCurve,
+        dropdownMenuEntries: const [
+          DropdownMenuEntry(value: 'bounceIn', label: 'bounceIn'),
+          DropdownMenuEntry(value: 'bounceInOut', label: 'bounceInOut'),
+          DropdownMenuEntry(value: 'bounceOut', label: 'bounceOut'),
+          DropdownMenuEntry(value: 'ease', label: 'ease'),
+          DropdownMenuEntry(value: 'easeIn', label: 'easeIn'),
+          DropdownMenuEntry(value: 'easeInBack', label: 'easeInBack'),
+          DropdownMenuEntry(value: 'easeInCirc', label: 'easeInCirc'),
+          DropdownMenuEntry(value: 'easeInExpo', label: 'easeInExpo'),
+          DropdownMenuEntry(value: 'easeInOut', label: 'easeInOut'),
+          DropdownMenuEntry(
+            value: 'easeInOutBack',
+            label: 'easeInOutBack',
+          ),
+          DropdownMenuEntry(
+            value: 'easeInOutCirc',
+            label: 'easeInOutCirc',
+          ),
+          DropdownMenuEntry(
+            value: 'easeInOutExpo',
+            label: 'easeInOutExpo',
+          ),
+          DropdownMenuEntry(
+            value: 'easeInOutQuad',
+            label: 'easeInOutQuad',
+          ),
+          DropdownMenuEntry(
+            value: 'easeInOutQuart',
+            label: 'easeInOutQuart',
+          ),
+          DropdownMenuEntry(
+            value: 'easeInOutQuint',
+            label: 'easeInOutQuint',
+          ),
+          DropdownMenuEntry(
+            value: 'easeInOutSine',
+            label: 'easeInOutSine',
+          ),
+          DropdownMenuEntry(value: 'easeInQuad', label: 'easeInQuad'),
+          DropdownMenuEntry(value: 'easeInQuart', label: 'easeInQuart'),
+          DropdownMenuEntry(value: 'easeInQuint', label: 'easeInQuint'),
+          DropdownMenuEntry(value: 'easeInSine', label: 'easeInSine'),
+          DropdownMenuEntry(value: 'easeOut', label: 'easeOut'),
+          DropdownMenuEntry(value: 'easeOutBack', label: 'easeOutBack'),
+          DropdownMenuEntry(value: 'easeOutCirc', label: 'easeOutCirc'),
+          DropdownMenuEntry(value: 'easeOutExpo', label: 'easeOutExpo'),
+          DropdownMenuEntry(value: 'easeOutQuad', label: 'easeOutQuad'),
+          DropdownMenuEntry(value: 'easeOutQuart', label: 'easeOutQuart'),
+          DropdownMenuEntry(value: 'easeOutQuint', label: 'easeOutQuint'),
+          DropdownMenuEntry(value: 'easeOutSine', label: 'easeOutSine'),
+          DropdownMenuEntry(value: 'elasticIn', label: 'elasticIn'),
+          DropdownMenuEntry(value: 'elasticInOut', label: 'elasticInOut'),
+          DropdownMenuEntry(value: 'elasticOut', label: 'elasticOut'),
+          DropdownMenuEntry(value: 'linear', label: 'linear'),
+        ],
+      );
+
+  /// The row with the buttons for adding, expanding, and collapsing
   Row _buttonRow() => Row(
         children: [
           FloatingActionButton(
@@ -387,8 +260,14 @@ class _MyAppState extends State<MyApp> {
         ],
       );
 
+  /// The segmented button for toggling the display style
   Widget _styleToggle() => SegmentedButton<DisplayStyle>(
         segments: const [
+          ButtonSegment(
+            value: DisplayStyle.crazy,
+            label: Text('Crazy'),
+            icon: Icon(Icons.style),
+          ),
           ButtonSegment(
             value: DisplayStyle.modern,
             label: Text('Modern'),
@@ -408,11 +287,12 @@ class _MyAppState extends State<MyApp> {
         },
       );
 
+  /// The tree view widget
   TreeView<FileSystemElement> _treeView() => TreeView(
         onSelectionChanged: (node) => setState(() => _selectedNode = node),
         key: treeViewKey,
         animationDuration: Duration(milliseconds: _animationDuration),
-        animationCurve: _getAnimationCurve(),
+        animationCurve: getAnimationCurve(_selectedCurve),
         builder: (
           context,
           node,
@@ -431,41 +311,7 @@ class _MyAppState extends State<MyApp> {
         ),
       );
 
-  Curve _getAnimationCurve() => switch (_selectedCurve) {
-        ('bounceIn') => Curves.bounceIn,
-        ('bounceInOut') => Curves.bounceInOut,
-        ('bounceOut') => Curves.bounceOut,
-        ('easeInCirc') => Curves.easeInCirc,
-        ('easeInOutExpo') => Curves.easeInOutExpo,
-        ('elasticInOut') => Curves.elasticInOut,
-        ('easeInOut') => Curves.easeInOut,
-        ('easeOutCirc') => Curves.easeOutCirc,
-        ('elasticOut') => Curves.elasticOut,
-        ('elasticIn') => Curves.elasticIn,
-        ('easeIn') => Curves.easeIn,
-        ('ease') => Curves.ease,
-        ('easeInBack') => Curves.easeInBack,
-        ('easeOutBack') => Curves.easeOutBack,
-        ('easeInOutBack') => Curves.easeInOutBack,
-        ('easeInSine') => Curves.easeInSine,
-        ('easeOutSine') => Curves.easeOutSine,
-        ('easeInOutSine') => Curves.easeInOutSine,
-        ('easeInQuad') => Curves.easeInQuad,
-        ('easeOutQuad') => Curves.easeOutQuad,
-        ('easeInOutQuad') => Curves.easeInOutQuad,
-        ('easeInQuart') => Curves.easeInQuart,
-        ('easeOutQuart') => Curves.easeOutQuart,
-        ('easeInOutQuart') => Curves.easeInOutQuart,
-        ('easeInQuint') => Curves.easeInQuint,
-        ('easeOutQuint') => Curves.easeOutQuint,
-        ('easeInOutQuint') => Curves.easeInOutQuint,
-        ('easeInExpo') => Curves.easeInExpo,
-        ('easeOutExpo') => Curves.easeOutExpo,
-        ('linear') => Curves.linear,
-        _ => Curves.easeInOut,
-      };
-
-  /// Renders file nodes with different styles based on the selected display 
+  /// Renders file nodes with different styles based on the selected display
   /// style
   Widget _file(
     void Function(TreeNode<FileSystemElement> node) select,
@@ -476,8 +322,10 @@ class _MyAppState extends State<MyApp> {
       switch (_displayStyle) {
         DisplayStyle.modern => _modernFile(select, node, isSelected, context),
         DisplayStyle.retro => _retroFile(select, node, isSelected, context),
+        DisplayStyle.crazy => _crazyFile(select, node, isSelected, context),
       };
 
+  /// The modern file node
   Widget _modernFile(
     void Function(TreeNode<FileSystemElement> node) select,
     TreeNode<FileSystemElement> node,
@@ -512,11 +360,11 @@ class _MyAppState extends State<MyApp> {
                       color: Theme.of(context)
                           .colorScheme
                           .primary
-                          .withValues(alpha:0.3),
+                          .withValues(alpha: 0.3),
                       blurRadius: 8,
                       offset: const Offset(0, 2),
-                    )
-                  ,]
+                    ),
+                  ]
                 : null,
           ),
           child: Padding(
@@ -567,6 +415,7 @@ class _MyAppState extends State<MyApp> {
         ),
       );
 
+  /// The retro file node
   Widget _retroFile(
     void Function(TreeNode<FileSystemElement> node) select,
     TreeNode<FileSystemElement> node,
@@ -580,7 +429,7 @@ class _MyAppState extends State<MyApp> {
           decoration: BoxDecoration(
             color: isSelected ? Colors.green.withValues(alpha: 0.2) : null,
             border: Border.all(
-              color: Colors.green.withValues(alpha:  0.5),
+              color: Colors.green.withValues(alpha: 0.5),
             ),
           ),
           child: Padding(
@@ -617,7 +466,7 @@ class _MyAppState extends State<MyApp> {
         ),
       );
 
-  /// Renders folder nodes with different styles based on the selected display 
+  /// Renders folder nodes with different styles based on the selected display
   /// style
   Widget _folder(
     Animation<double> expansionAnimation,
@@ -626,8 +475,86 @@ class _MyAppState extends State<MyApp> {
       switch (_displayStyle) {
         DisplayStyle.modern => _modernFolder(expansionAnimation, node),
         DisplayStyle.retro => _retroFolder(expansionAnimation, node),
+        DisplayStyle.crazy => _crazyfolder(expansionAnimation, node),
       };
 
+  /// Renders file nodes
+  InkWell _crazyFile(
+    void Function(TreeNode<FileSystemElement> node) select,
+    TreeNode<FileSystemElement> node,
+    bool isSelected,
+    BuildContext context,
+  ) =>
+      InkWell(
+        onTap: () => select(node),
+        // ignore: use_decorated_box
+        child: Container(
+          margin: const EdgeInsets.symmetric(horizontal: 8),
+          decoration: BoxDecoration(
+            gradient: isSelected
+                ? LinearGradient(
+                    colors: [
+                      Theme.of(context)
+                          .colorScheme
+                          .primary
+                          .withValues(alpha: 0.3),
+                      Theme.of(context)
+                          .colorScheme
+                          .primary
+                          .withValues(alpha: 0.1),
+                    ],
+                    begin: Alignment.topLeft,
+                    end: Alignment.bottomRight,
+                  )
+                : null,
+            borderRadius: BorderRadius.circular(4),
+          ),
+          child: Padding(
+            padding: const EdgeInsets.all(8),
+            child: Row(
+              children: [
+                imageAsset(
+                  switch (path.extension(node.data.name).toLowerCase()) {
+                    ('.mp3') => 'assets/images/music.png',
+                    ('.py') => 'assets/images/python.png',
+                    ('.jpg') => 'assets/images/image.png',
+                    ('.png') => 'assets/images/image.png',
+                    ('.dart') => 'assets/images/dart.png',
+                    ('.json') => 'assets/images/json.png',
+                    (_) => 'assets/images/file.png'
+                  },
+                  width: 32,
+                  height: 32,
+                ),
+                const SizedBox(width: 16),
+                Text(node.data.name),
+              ],
+            ),
+          ),
+        ),
+      );
+
+  /// Renders tree nodes
+  Row _crazyfolder(
+    Animation<double> expansionAnimation,
+    TreeNode<FileSystemElement> node,
+  ) =>
+      Row(
+        children: [
+          RotationTransition(
+            turns: expansionAnimation,
+            child: imageAsset(
+              'assets/images/folder.png',
+              width: 32,
+              height: 32,
+            ),
+          ),
+          const SizedBox(width: 16),
+          Text(node.data.name),
+        ],
+      );
+
+  /// The modern folder node
   Widget _modernFolder(
     Animation<double> expansionAnimation,
     TreeNode<FileSystemElement> node,
@@ -637,8 +564,8 @@ class _MyAppState extends State<MyApp> {
         decoration: BoxDecoration(
           gradient: LinearGradient(
             colors: [
-              Colors.blue.withValues(alpha:0.2),
-              Colors.purple.withValues(alpha:0.2),
+              Colors.blue.withValues(alpha: 0.2),
+              Colors.purple.withValues(alpha: 0.2),
             ],
             begin: Alignment.topLeft,
             end: Alignment.bottomRight,
@@ -654,7 +581,7 @@ class _MyAppState extends State<MyApp> {
                 child: Container(
                   padding: const EdgeInsets.all(8),
                   decoration: BoxDecoration(
-                    color: Colors.blue.withValues(alpha:0.2),
+                    color: Colors.blue.withValues(alpha: 0.2),
                     borderRadius: BorderRadius.circular(8),
                   ),
                   child: const Icon(
@@ -688,6 +615,7 @@ class _MyAppState extends State<MyApp> {
         ),
       );
 
+  /// The retro folder node
   Widget _retroFolder(
     Animation<double> expansionAnimation,
     TreeNode<FileSystemElement> node,
@@ -696,7 +624,7 @@ class _MyAppState extends State<MyApp> {
         margin: const EdgeInsets.symmetric(horizontal: 8, vertical: 2),
         decoration: BoxDecoration(
           border: Border.all(
-            color: Colors.green.withValues(alpha:0.5),
+            color: Colors.green.withValues(alpha: 0.5),
           ),
         ),
         child: Padding(
